fix: review of Tinh
