--- model/courseDetails.ts
@@ -1,6 +1,0 @@
-export interface ICourseDetails {
-    id: number;
-    image: JSX.Element;
-    title1: string;
-    title2: string;
-}
