--- model/decriptionCourse.ts
@@ -1,7 +1,0 @@
-export interface IDescriptionCourse {
-    title: string;
-    description: string;
-    cost: string;
-    count: string;
-    image: any;
-}
