+++ styles/common/shadow.css
@@ -1,3 +1,12 @@
 .custom-shadow {
     box-shadow: 3px 3px 0px 0px #7fcffc;
 }
+
+
+
+
+
+
+
+
+
