+++ styles/globals.css
@@ -63,7 +63,6 @@ input[type="text"]:not(:placeholder-shown) ~ .input-text {
     cursor: pointer;
 }
 
-
 .triangle {
     width: 0;
     height: 0;
@@ -93,21 +92,38 @@ input[type="text"]:not(:placeholder-shown) ~ .input-text {
 .video_container:hover .closeButton {
     opacity: 100%;
 }
-.backgroud-shadow {
-    /* box-shadow: rgba(0, 0, 0, 0.25) 0px 54px 55px, rgba(0, 0, 0, 0.12) 0px -12px 30px, rgba(0, 0, 0, 0.12) 0px 4px 6px, rgba(0, 0, 0, 0.17) 0px 12px 13px, rgba(0, 0, 0, 0.09) 0px -3px 5px; */
-    /* box-shadow: rgba(50, 50, 93, 0.25) 0px 50px 100px -20px, rgba(0, 0, 0, 0.3) 0px 30px 60px -30px, rgba(10, 37, 64, 0.35) 0px -2px 6px 0px inset; */
-    /* box-shadow: rgb(85, 91, 255) 0px 0px 0px 3px, rgb(31, 193, 27) 0px 0px 0px 6px, rgb(255, 217, 19) 0px 0px 0px 9px, rgb(255, 156, 85) 0px 0px 0px 12px, rgb(255, 85, 85) 0px 0px 0px 15px; */
-}
 
 .dgu-course-card {
-  transition:  0.2s all;
-  box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+    transition: 0.2s all;
+    box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
 }
 
 .dgu-course-card:hover {
-  transform:  scale(1.03);
-  background-color: white;
-  z-index: 5;
+    transform: scale(1.03);
+    background-color: white;
+    z-index: 5;
 }
 
+/* ---------------------scroll-thin-course------------------------- */
+.scroll-thin-course::-webkit-scrollbar {
+    width: 5px;
+}
 
+/* Track */
+.scroll-thin-course::-webkit-scrollbar-track {
+    background: #d7d7d7;
+    border-radius: 5px;
+}
+
+/* Handle */
+.scroll-thin-course::-webkit-scrollbar-thumb {
+    background: #b7b7b7;
+    border-radius: 5px;
+    padding: 2px;
+}
+
+/* Handle on hover */
+.scroll-thin-course::-webkit-scrollbar-thumb:hover {
+    background: rgb(146, 146, 146);
+    cursor: pointer;
+}
