--- styles/tailwindConfig.css
@@ -1,8 +1,0 @@
-.box-sizing-initial {
-    box-sizing: initial;
-}
-
-.space-between {
-    align-content: space-between;
-}
-
